+++ Packages/UGF.Types/Runtime/TypesUtility.cs
@@ -21,14 +21,14 @@ namespace UGF.Types.Runtime
         public static void GetTypes<T>(ITypeProvider<T> provider)
         {
             if (provider == null) throw new ArgumentNullException(nameof(provider));
-            
+
             var types = new List<Type>();
-            
+
             AssemblyUtility.GetBrowsableTypes(types, typeof(TypeIdentifierAttributeBase));
-            
+
             AddTypes(provider, types);
         }
-        
+
         /// <summary>
         /// Adds all found types marked with identifier attribute and register them into the specified provider from the specified assembly.
         /// <para>
@@ -41,11 +41,11 @@ namespace UGF.Types.Runtime
         {
             if (provider == null) throw new ArgumentNullException(nameof(provider));
             if (assembly == null) throw new ArgumentNullException(nameof(assembly));
-            
+
             var types = new List<Type>();
-            
+
             AssemblyUtility.GetBrowsableTypes(types, assembly, typeof(TypeIdentifierAttributeBase));
-            
+
             AddTypes(provider, types);
         }
 
@@ -58,7 +58,7 @@ namespace UGF.Types.Runtime
         {
             if (provider == null) throw new ArgumentNullException(nameof(provider));
             if (types == null) throw new ArgumentNullException(nameof(types));
-            
+
             for (int i = 0; i < types.Count; i++)
             {
                 Type type = types[i];
@@ -72,6 +72,89 @@ namespace UGF.Types.Runtime
                     Debug.LogWarning($"Can not add type to provider, cause '{typeof(ITypeIdentifierAttribute<T>)}' not found at specified type: '{type}'.");
                 }
             }
+        }
+
+        /// <summary>
+        /// Collects all types into the specified collection that match by specified func condition, if presents.
+        /// <para>
+        /// If validate func does not specified, will add all types.
+        /// </para>
+        /// </summary>
+        /// <param name="types">The collection to add types.</param>
+        /// <param name="validate">The function to validate type.</param>
+        public static void CollectTypes(ICollection<Type> types, Func<Type, bool> validate = null)
+        {
+            if (types == null) throw new ArgumentNullException(nameof(types));
+
+            Assembly[] assemblies = AppDomain.CurrentDomain.GetAssemblies();
+
+            for (int i = 0; i < assemblies.Length; i++)
+            {
+                CollectTypes(types, assemblies[i], validate);
+            }
+        }
+
+        /// <summary>
+        /// Collects all types the specified assembly into the specified collection, that match by specified func condition, if presents.
+        /// <para>
+        /// If validate func does not specified, will add all types from the assembly.
+        /// </para>
+        /// </summary>
+        /// <param name="types">The collection to add types.</param>
+        /// <param name="assembly">The assembly to gather types.</param>
+        /// <param name="validate">The function to validate type.</param>
+        public static void CollectTypes(ICollection<Type> types, Assembly assembly, Func<Type, bool> validate = null)
+        {
+            if (types == null) throw new ArgumentNullException(nameof(types));
+            if (assembly == null) throw new ArgumentNullException(nameof(types));
+
+            Type[] assemblyTypes = assembly.GetTypes();
+
+            for (int i = 0; i < assemblyTypes.Length; i++)
+            {
+                Type type = assemblyTypes[i];
+
+                if (validate == null || validate(type))
+                {
+                    types.Add(type);
+                }
+            }
+        }
+
+        /// <summary>
+        /// Tries to get type identifier from the specified type that contains identifier attribute.
+        /// </summary>
+        /// <param name="type">The target type.</param>
+        /// <param name="identifier">The found identifier.</param>
+        public static bool TryGetIdentifierFromType<T>(Type type, out T identifier)
+        {
+            if (type.GetCustomAttribute<TypeIdentifierAttributeBase>() is ITypeIdentifierAttribute<T> attribute)
+            {
+                identifier = attribute.Identifier;
+                return true;
+            }
+
+            identifier = default;
+            return false;
+        }
+
+        /// <summary>
+        /// Tries to get type identifier from the specified type that contains identifier attribute.
+        /// </summary>
+        /// <param name="type">The target type.</param>
+        /// <param name="identifier">The found identifier.</param>
+        public static bool TryGetIdentifierFromType(Type type, out object identifier)
+        {
+            var attribute = type.GetCustomAttribute<TypeIdentifierAttributeBase>();
+
+            if (attribute != null)
+            {
+                identifier = attribute.GetIdentifier();
+                return true;
+            }
+
+            identifier = null;
+            return false;
         }
     }
 }