+++ changelog.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Milestone](https://github.com/unity-game-framework/ugf-types/milestone/0?closed=1)
 
 ### Added
-- Nothing.
+- `TypesUtility.TryGetIdentifierFromType` to retrieve type identifier from attribute directly.
 
 ### Changed
 - Nothing.