+++ changelog.md
@@ -9,8 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Milestone](https://github.com/unity-game-framework/ugf-types/milestone/0?closed=1)
 
 ### Added
-- `CollectTypes` to collect types with validation.
-- `TypesDropdown` - editor advanced dropdown to display types selection menu.
+- `TypesUtility.CollectTypes` to collect types with validation.
+- `TypesEditorGUIUtility.GetTypesDropdown` to create an advanced editor dropdown that displays types selection menu.
 
 ### Changed
 - Nothing.