+++ changelog.md
@@ -26,6 +26,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Security
 - Nothing.
 
+## 1.1.0 - 2019-04-14
+- [Commits](https://github.com/unity-game-framework/ugf-types/compare/1.0.0...1.1.0)
+- [Milestone](https://github.com/unity-game-framework/ugf-types/milestone/2?closed=1)
+
+### Added
+- `TypesUtility.CollectTypes` to collect types with validation.
+- `TypesEditorGUIUtility.GetTypesDropdown` to create an advanced editor dropdown that displays types selection menu.
+- `TypesUtility.TryGetIdentifierFromType` to retrieve type identifier from attribute directly.
+
 ## 1.0.0 - 2019-03-24
 - [Commits](https://github.com/unity-game-framework/ugf-types/compare/93305d0...1.0.0)
 - [Milestone](https://github.com/unity-game-framework/ugf-types/milestone/1?closed=1)